storage: add authId, authDomain to virDomainDiskDef
Add additional fields to let you specify the how to authenticate with a
network disk type.  The authId is the name to authenticate as, and the
authDomain optionally describes the domain that user exists in.  The latter
allows us to locate a secret in using the libvirt secrets API, as the user
is may not unique if libvirt is talking to multiple backend clusters.

Signed-off-by: Sage Weil <sage@newdream.net>